fix : dto 필드 수정

+++ src/main/java/live/ioteatime/frontservice/dto/Alert.java
@@ -12,7 +12,7 @@ public class Alert {
     private String sensorName;
     private String place;
     private String type;
-    private int time;
+    private long time;
     private double outlierValue;
     private int organizationId;
 }
